+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 
 ## 0.3.2 - 2023-10-09
 - Cleaned up code and some potential Bugs ruff/pyright
+- Added movement Above/Below and updated others to use it
 
 ## 0.3.1 - 2023-10-09
 - Added candles_timerange to auto remove older than N candles